+++ test/plugin/add_metadata_test.rb
@@ -4,119 +4,100 @@ require "shrine/plugins/add_metadata"
 describe Shrine::Plugins::AddMetadata do
   before do
     @uploader = uploader { plugin :add_metadata }
+    @shrine = @uploader.class
   end
 
   describe "add_metadata" do
     describe "with argument" do
       it "adds declared metadata" do
-        @uploader.class.add_metadata(:custom) { |io, context| "value" }
-        uploaded_file = @uploader.upload(fakeio)
-        assert_equal "value", uploaded_file.metadata.fetch("custom")
-      end
-
-      it "executes inside uploader and forwards correct arguments" do
-        @uploader.class.add_metadata(:custom) do |io, context|
-          raise unless self.is_a?(Shrine)
-          raise unless io.respond_to?(:read)
-          raise unless context.is_a?(Hash) && context.key?(:foo)
-        end
-        @uploader.upload(fakeio, foo: "bar")
+        @shrine.add_metadata(:custom) { |io, context| "value" }
+        metadata = @uploader.extract_metadata(fakeio)
+        assert_equal "value", metadata.fetch("custom")
+        assert_instance_of Integer, metadata["size"]
       end
 
       it "allows returning nil" do
-        @uploader.class.add_metadata(:custom) { nil }
-        uploaded_file = @uploader.upload(fakeio)
-        refute uploaded_file.metadata.key?("custom")
-      end
-
-      it "rewinds the IO after extracting metadata" do
-        @uploader.class.add_metadata(:custom) { |io, context| io.read }
-        uploaded_file = @uploader.upload(fakeio("file"))
-        assert_equal "file", uploaded_file.read
+        @shrine.add_metadata(:custom) { nil }
+        metadata = @uploader.extract_metadata(fakeio)
+        assert_nil metadata.fetch("custom")
       end
 
       it "adds the metadata method to UploadedFile" do
-        @uploader.class.add_metadata(:custom) { |io, context| "value" }
+        @shrine.add_metadata(:custom) { |io, context| "value" }
         uploaded_file = @uploader.upload(fakeio)
         assert_equal "value", uploaded_file.custom
-      end
-
-      it "includes existing metadata in context" do
-        @uploader.class.add_metadata(:extracted_size) { |io, context| context[:metadata]["size"] }
-        @uploader.class.add_metadata(:extracted_content_type) { |io, context| context[:metadata]["mime_type"] }
-        @uploader.class.add_metadata(:extracted_filename) { |io, context| context[:metadata]["filename"] }
-        @uploader.class.add_metadata(:we_added) { |io, context| "added" }
-        @uploader.class.add_metadata(:extracted_we_added) { |io, context| context[:metadata]["we_added"] }
-
-        uploaded_file = @uploader.upload(fakeio(filename: "supplied_filename.txt", content_type: "text/plain"))
-
-        assert_equal "added", uploaded_file.extracted_we_added
-        assert_equal "supplied_filename.txt", uploaded_file.metadata["extracted_filename"]
-        assert_equal "text/plain", uploaded_file.metadata["extracted_content_type"]
-        assert_equal uploaded_file.size, uploaded_file.metadata["extracted_size"]
       end
     end
 
     describe "without argument" do
       it "adds declared metadata" do
-        @uploader.class.add_metadata { Hash["custom" => "value"] }
-        uploaded_file = @uploader.upload(fakeio)
-        assert_equal "value", uploaded_file.metadata.fetch("custom")
+        @shrine.add_metadata { Hash["custom" => "value"] }
+        metadata = @uploader.extract_metadata(fakeio)
+        assert_equal "value", metadata.fetch("custom")
+        assert_instance_of Integer, metadata["size"]
       end
 
       it "accepts symbol keys" do
-        @uploader.class.add_metadata { Hash[custom: "value"] }
-        uploaded_file = @uploader.upload(fakeio)
-        assert_equal "value", uploaded_file.metadata.fetch("custom")
-      end
-
-      it "executes inside uploader and forwards correct arguments" do
-        @uploader.class.add_metadata do |io, context|
-          raise unless self.is_a?(Shrine)
-          raise unless io.respond_to?(:read)
-          raise unless context.is_a?(Hash) && context.key?(:foo)
-        end
-        @uploader.upload(fakeio, foo: "bar")
+        @shrine.add_metadata { Hash[custom: "value"] }
+        metadata = @uploader.extract_metadata(fakeio)
+        assert_equal "value", metadata.fetch("custom")
       end
 
       it "allows returning nil" do
-        @uploader.class.add_metadata { nil }
-        uploaded_file = @uploader.upload(fakeio)
+        @shrine.add_metadata { nil }
+        metadata = @uploader.extract_metadata(fakeio)
+        assert_equal %w[filename size mime_type], metadata.keys
       end
+    end
 
-      it "rewinds the IO after extracting metadata" do
-        @uploader.class.add_metadata { |io, context| io.read; nil }
-        uploaded_file = @uploader.upload(fakeio("file"))
-        assert_equal "file", uploaded_file.read
+    it "executes inside uploader and forwards correct arguments" do
+      minitest = self
+      input = fakeio
+      @shrine.add_metadata(:custom) do |io, context|
+        minitest.assert_kind_of Shrine, self
+        minitest.assert_equal input, io
+        minitest.assert_instance_of Hash, context
+        minitest.assert_equal "bar", context[:foo]
+        nil
       end
+      @uploader.extract_metadata(input, foo: "bar")
+    end
 
-      it "includes existing metadata in context" do
-        @uploader.class.add_metadata(:we_added) { |io, context| "added" }
-        @uploader.class.add_metadata do |io, context|
-          { extracted_existing_metadata: context[:metadata].dup }
-        end
+    it "rewinds the IO between metadata blocks" do
+      @shrine.add_metadata(:foo) { |io| io.read }
+      @shrine.add_metadata(:bar) { |io| io.read }
+      metadata = @uploader.extract_metadata(input = fakeio("file"))
+      assert_equal "file", metadata["foo"]
+      assert_equal "file", metadata["bar"]
+      assert_equal "file", input.read
+    end
 
-        uploaded_file = @uploader.upload(fakeio(filename: "supplied_filename.txt", content_type: "text/plain"))
-        extracted_existing_metadata = uploaded_file.metadata["extracted_existing_metadata"]
+    it "includes existing metadata in context" do
+      @shrine.add_metadata(:extracted_size) { |io, context| context[:metadata]["size"] }
+      @shrine.add_metadata(:extracted_content_type) { |io, context| context[:metadata]["mime_type"] }
+      @shrine.add_metadata(:extracted_filename) { |io, context| context[:metadata]["filename"] }
+      @shrine.add_metadata(:we_added) { |io, context| "added" }
+      @shrine.add_metadata(:extracted_we_added) { |io, context| context[:metadata]["we_added"] }
 
-        assert_equal "added", extracted_existing_metadata["we_added"]
-        assert_equal "supplied_filename.txt", extracted_existing_metadata["filename"]
-        assert_equal "text/plain", extracted_existing_metadata["mime_type"]
-        assert_equal uploaded_file.size, extracted_existing_metadata["size"]
-      end
+      metadata = @uploader.extract_metadata(fakeio(filename: "supplied_filename.txt", content_type: "text/plain"))
+
+      assert_equal "added", metadata["extracted_we_added"]
+      assert_equal "supplied_filename.txt", metadata["extracted_filename"]
+      assert_equal "text/plain", metadata["extracted_content_type"]
+      assert_equal metadata["size"], metadata["extracted_size"]
     end
   end
 
   describe "metadata_method" do
     it "defines a reader on the uploaded file which returns the metadata" do
-      @uploader.class.add_metadata { Hash["custom" => "value"] }
-      @uploader.class.metadata_method :custom
+      @shrine.add_metadata { Hash["custom" => "value"] }
+      @shrine.metadata_method :custom
       uploaded_file = @uploader.upload(fakeio)
       assert_equal "value", uploaded_file.custom
     end
 
     it "returns nil if metadata is missing" do
-      @uploader.class.metadata_method :custom
+      @shrine.metadata_method :custom
       uploaded_file = @uploader.upload(fakeio)
       assert_nil uploaded_file.custom
     end